requirements.txt
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.4.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
attrs==24.3.0
av==14.0.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
comtypes==1.4.6
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
dashscope==1.21.0
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator==4.4.2
deepdiff==8.0.1
distro==1.9.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.13.1
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.2.0
gcloud==0.18.3
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-vision==3.7.4
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
gradio==5.12.0
gradio_client==1.5.4
grpcio==1.65.5
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.27.1
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1721856510709/work
intel-openmp==2021.4.0
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1719582682925/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
jiter==0.5.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///D:/bld/jupyter_core_1710257313664/work
jwcrypto==1.5.6
kiwisolver==1.4.5
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.4.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
oauth2client==4.1.3
openai==1.41.0
openai-whisper==20231117
opencv-python==4.10.0.84
orderly-set==5.2.2
orjson==3.10.14
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
proglog==0.1.10
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
propcache==0.2.1
proto-plus==1.24.0
protobuf==4.25.4
psutil @ file:///C:/ci_311_rebuilds/psutil_1679005906571/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
Pyrebase4==4.8.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-jwt==4.1.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2024.1
pywin32==305.1
pywinauto==0.6.8
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1666828541352/work
qwen-vl-utils==0.0.8
regex==2024.7.24
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.4
rsa==4.9
ruff==0.9.1
safehttpx==0.1.6
safetensors==0.4.4
scipy==1.14.0
screeninfo==0.8.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.1
soupsieve==2.6
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.41.3
sympy==1.12
tbb==2021.11.0
termcolor==2.4.0
textdistance==4.6.3
tiktoken==0.7.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.3.1+cpu
torchaudio==2.3.1+cpu
torchvision==0.18.1+cpu
tornado @ file:///D:/bld/tornado_1666788735597/work
tqdm==4.66.5
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.48.0
typer==0.15.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
uiautomation==2.0.20
ultralytics==8.2.53
ultralytics-thop==2.0.5
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.3
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1723591248676/work
